skeleton done

+++ benchmarks/src/config.rs
@@ -17,9 +17,9 @@ pub struct Config {
 
     /// how many transactions to send
     #[clap(short, long, default_value = "1000")]
-    pub n_transactions: u32,
+    pub n_transactions: u64,
 
     /// how long to run the benchmark for (in seconds)
     #[clap(short, long, default_value = "10")]
-    pub time: u32,
+    pub time: u64,
 }
